CustomerSupportAgent/requirements.txt
langgraph
langchain-core
langchain_google_genai
python-dotenv
ipython
streamlit